Use hawkular-commons 0.9.1, update CORS
It means splitting cors filtering into a request filter and a response filter.
Validation is processed during request, adding headers dureing response.
