+++ hawkular-inventory-rest-api/src/main/java/org/hawkular/inventory/rest/interceptors/CorsResponseFilter.java
@@ -1,0 +1,57 @@
+/*
+ * Copyright 2015-2016 Red Hat, Inc. and/or its affiliates
+ * and other contributors as indicated by the @author tags.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.hawkular.inventory.rest.interceptors;
+
+import java.io.IOException;
+
+import javax.annotation.PostConstruct;
+import javax.inject.Inject;
+import javax.ws.rs.container.ContainerRequestContext;
+import javax.ws.rs.container.ContainerResponseContext;
+import javax.ws.rs.container.ContainerResponseFilter;
+import javax.ws.rs.ext.Provider;
+
+import org.hawkular.inventory.api.Configuration;
+import org.hawkular.jaxrs.filter.cors.CorsFilters;
+import org.jboss.resteasy.annotations.interception.ServerInterceptor;
+
+/**
+ * @author Joel Takvorian
+ */
+@Provider
+@ServerInterceptor
+public class CorsResponseFilter implements ContainerResponseFilter {
+
+    @Inject
+    private Configuration configuration;
+
+    private String extraAllowedHeaders;
+
+    @PostConstruct
+    protected void init() {
+        extraAllowedHeaders = "authorization";
+        String headers = configuration.getProperty(CorsProperties.ALLOWED_CORS_ACCESS_CONTROL_ALLOW_HEADERS, null);
+        if (headers != null) {
+            extraAllowedHeaders += "," + headers;
+        }
+    }
+
+    @Override public void filter(ContainerRequestContext requestContext, ContainerResponseContext responseContext)
+            throws IOException {
+        CorsFilters.filterResponse(requestContext, responseContext, extraAllowedHeaders);
+    }
+}
